simplify logic test to new file
