+++ Wire-iOS Tests/ConversationAvatarViewTests.swift
@@ -35,84 +35,6 @@ extension Array where Element: ZMUser {
     }
 }
 
-protocol CoreDataFixed {
-    var coreDataFixture: CoreDataFixture! { get }
-    var otherUser: ZMUser! { get }
-    var selfUser: ZMUser! { get }
-
-    func createGroupConversation() -> ZMConversation
-}
-
-extension CoreDataFixed {
-    var otherUser: ZMUser! {
-        return coreDataFixture.otherUser
-    }
-
-    var selfUser: ZMUser! {
-        return coreDataFixture.selfUser
-    }
-
-    func createGroupConversation() -> ZMConversation {
-        return coreDataFixture.createGroupConversation()
-    }
-}
-
-final class ConversationAvatarViewModeTests: XCTestCase, CoreDataFixed {
-    var sut: ConversationAvatarView!
-
-    var coreDataFixture: CoreDataFixture!
-
-    override func setUp() {
-        super.setUp()
-        coreDataFixture = CoreDataFixture()
-        sut = ConversationAvatarView()
-    }
-
-    override func tearDown() {
-        sut = nil
-        coreDataFixture = nil
-        super.tearDown()
-    }
-
-    func testThatModeIsOneWhenGroupConversationWithOneServiceUser() {
-        // GIVEN
-        otherUser.serviceIdentifier = "serviceIdentifier"
-        otherUser.providerIdentifier = "providerIdentifier"
-        XCTAssert(otherUser.isServiceUser)
-
-        let conversation = createGroupConversation()
-
-        // WHEN
-        sut.configure(context: .conversation(conversation: conversation))
-
-        // THEN
-        XCTAssertEqual(sut.mode, .one(serviceUser: true))
-    }
-
-    func testThatModeIsFourWhenGroupConversationWithOneUser() {
-        // GIVEN
-        let conversation = createGroupConversation()
-
-        // WHEN
-        sut.configure(context: .conversation(conversation: conversation))
-
-        // THEN
-        XCTAssertEqual(sut.mode, .four)
-    }
-
-    func testThatModeIsNoneWhenGroupConversationIsEmpty() {
-        // GIVEN
-        let conversation = createGroupConversation()
-        conversation.internalRemoveParticipants([otherUser!], sender: selfUser)
-
-        // WHEN
-        sut.configure(context: .conversation(conversation: conversation))
-
-        // THEN
-        XCTAssertEqual(sut.mode, .none)
-    }
-}
-
 final class ConversationAvatarViewTests: CoreDataSnapshotTestCase {
 
     var sut: ConversationAvatarView!
